+++ sample_projects/dependency_issue_65/default.nix
@@ -1,0 +1,6 @@
+{ pkgs? import ../../nixpkgs.nix { config = {}; }
+, generatedBuild ? ./Cargo.nix { } }:
+
+let cargo_nix = pkgs.callPackage generatedBuild {};
+    bin = cargo_nix.rootCrate.build.override { features = ["sqlite"]; };
+in bin